Move types to new crate
