+++ Cargo.toml
@@ -2,6 +2,7 @@
 default-members = ["kube"]
 members = [
   "kube",
+  "kube-core",
   "kube-derive",
   "kube-runtime",
