+++ kube-core/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod api_resource;
+pub mod gvk;
+pub mod resource;
