+++ kube-derive/Cargo.toml
@@ -28,7 +28,7 @@ schema = []
 [dev-dependencies]
 serde = { version = "1.0.118", features = ["derive"] }
 serde_yaml = "0.8.17"
-kube = { path = "../kube", version = "<1.0.0, >=0.51.0" }
+kube-core = { path = "../kube-core" }
 k8s-openapi = { version = "0.11.0", default-features = false, features = ["v1_20"] }
 schemars = { version = "0.8.0", features = ["chrono"] }
 chrono = "0.4.19"
