+++ kube/Cargo.toml
@@ -52,6 +52,7 @@ bytes = "1.0.0"
 tokio = { version = "1.0.1", features = ["time", "signal", "sync"] }
 static_assertions = "1.1.0"
 kube-derive = { path = "../kube-derive", version = "^0.53.0", optional = true }
+kube-core = { path = "../kube-core", version = "^0.53.0" }
 jsonpath_lib = "0.2.6"
 tokio-util = { version = "0.6.0", features = ["io", "codec"] }
 json-patch = { version = "0.2.6", optional = true }
